Move new SIMD API to rten-simd crate root
After removing the old SIMD API, the new one can be moved to the crate root.
