+++ rten-simd/src/arch/aarch64.rs
@@ -16,7 +16,7 @@ use std::arch::aarch64::{
 };
 use std::mem::transmute;
 
-use crate::safe::{
+use crate::{
     Extend, FloatOps, Interleave, Isa, Mask, MaskOps, NarrowSaturate, NumOps, SignedIntOps, Simd,
 };
